📝 docs(LICENSE): Add MIT license to the project
This commit adds the MIT license to the project. The MIT license grants permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell the software. The license also includes a warranty disclaimer. The copyright notice and permission notice are included in all copies or substantial portions of the software.

Changes made:
- Added a new file `LICENSE` with the text of the MIT license.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>